refactor: 스타일 적용 훅 분리
